+++ .eslintrc.cjs
@@ -1,25 +1,36 @@
 // This is a workaround for https://github.com/eslint/eslint/issues/3458
-require("@rushstack/eslint-patch/modern-module-resolution");
+require('@rushstack/eslint-patch/modern-module-resolution');
+
+
 
 module.exports = {
-  "env": {
-    "es2021": true,
-    "node": true
+  env: {
+    es2021: true,
+    node: true,
   },
-  "extends": [
-    "plugin:@typescript-eslint/recommended-requiring-type-checking", // *1
-    "eslint-config-gev", // https://github.com/SrBrahma/eslint-config-gev
+  extends: [
+    'eslint-config-gev/js', // https://github.com/SrBrahma/eslint-config-gev
   ],
-  "parser": "@typescript-eslint/parser",
-  "parserOptions": {
-    "tsconfigRootDir": __dirname,   // *1
-    "project": ['./tsconfig.lint.json'], // *1
-    "ecmaVersion": 12,
-    "sourceType": "module",
+  parser: '@typescript-eslint/parser',
+  overrides: [ //* 2
+    {
+      files: ['*.ts', '*.tsx'],
+      extends: [
+        'plugin:@typescript-eslint/recommended-requiring-type-checking', // *1
+        'eslint-config-gev/react-native', // https://github.com/SrBrahma/eslint-config-gev
+      ],
+      parserOptions: {
+        tsconfigRootDir: __dirname, // *1
+        project: ['./tsconfig.lint.json'], // *1
+        ecmaVersion: 12,
+        sourceType: 'module',
+      },
+    },
+  ],
+  ignorePatterns: ['**/lib/**/*', '**/dist/**/*', '**/node_modules/**/*', '.eslintrc.cjs'],
+  rules: {
   },
-  "rules": {
-  }
 };
 
-// [*1] - Optional but improves the linting for TS:
-// https://github.com/typescript-eslint/typescript-eslint/blob/master/docs/getting-started/linting/TYPED_LINTING.md#getting-started---linting-with-type-information
+// [*1] - https://github.com/typescript-eslint/typescript-eslint/blob/master/docs/getting-started/linting/TYPED_LINTING.md#getting-started---linting-with-type-information
+// [*2] - https://stackoverflow.com/a/64488474