+++ semitemplates/expo/jest.setup.js
@@ -2,6 +2,7 @@
 require('react-native-gesture-handler/jestSetup');
 
 
+
 jest.mock('react-native-reanimated', () => {
   const Reanimated = require('react-native-reanimated/mock');